Senior Cloud Architect
Various projects in the context of SaaS solutions within the healthcare sector.
The overarching theme was modernizing various components to use cloud native technologies after a prior "lift and shift" to Azure.
In my role as architect for the department, I lead the technical design and implementation for teams of up to 14 people per team.